Ajuste Listagem categorias

+++ app/src/main/java/br/senac/pi/model/Categoria.kt
@@ -1,7 +1,7 @@
 package br.senac.pi.model
 
 data class Categoria(
-    var CATEGORIA_ID: Long,
+    var CATEGORIA_ID: Int,
     var CATEGORIA_NOME: String,
     var CATEGORIA_DESC: String,
     var CATEGORIA_ATIVO: Float,
